moving tests and PEPE 8 fixes in flatten.py
