+++ array/delete_nth.py
@@ -7,7 +7,6 @@ drop the next [1,2] since this would lead to 1 and 2 being in the result 3 times
 which leads to [1,2,3,1,2,3]
 """
 import collections
-import unittest
 
 
 # Time complexity O(n^2)
@@ -31,38 +30,3 @@ def delete_nth(array, n):
             counts[i] += 1
 
     return result
-
-
-class TestSuite(unittest.TestCase):
-
-    def test_delete_nth_naive(self):
-
-        self.assertListEqual(delete_nth_naive([20, 37, 20, 21, 37, 21, 21], n=1),
-                             [20, 37, 21])
-        self.assertListEqual(delete_nth_naive([1, 1, 3, 3, 7, 2, 2, 2, 2], n=3),
-                             [1, 1, 3, 3, 7, 2, 2, 2])
-        self.assertListEqual(delete_nth_naive([1, 2, 3, 1, 1, 2, 1, 2, 3, 3, 2, 4, 5, 3, 1], n=3),
-                             [1, 2, 3, 1, 1, 2, 2, 3, 3, 4, 5])
-        self.assertListEqual(delete_nth_naive([], n=5),
-                             [])
-        self.assertListEqual(delete_nth_naive([1, 2, 3, 1, 1, 2, 1, 2, 3, 3, 2, 4, 5, 3, 1], n=0),
-                             [])
-
-    def test_delete_nth(self):
-
-        self.assertListEqual(delete_nth([20, 37, 20, 21, 37, 21, 21], n=1),
-                             [20, 37, 21])
-        self.assertListEqual(delete_nth([1, 1, 3, 3, 7, 2, 2, 2, 2], n=3),
-                             [1, 1, 3, 3, 7, 2, 2, 2])
-        self.assertListEqual(delete_nth([1, 2, 3, 1, 1, 2, 1, 2, 3, 3, 2, 4, 5, 3, 1], n=3),
-                             [1, 2, 3, 1, 1, 2, 2, 3, 3, 4, 5])
-        self.assertListEqual(delete_nth([], n=5),
-                             [])
-        self.assertListEqual(delete_nth([1, 2, 3, 1, 1, 2, 1, 2, 3, 3, 2, 4, 5, 3, 1], n=0),
-                             [])
-
-
-if __name__ == '__main__':
-
-    unittest.main()
-
